requirements.txt
click==5.1
colorama==0.3.3
coreapi==2.0.8
Django==1.8
django-allauth==0.28.0
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-messages==0.5.3
django-postman==3.4.0
django-rest-swagger==2.0.7
django-reversion==2.0.6
django-user-accounts==2.0.0
djangorestframework==3.4.7
itypes==1.1.0
Markdown==2.6.7
MySQL-python==1.2.5
oauthlib==2.0.0
openapi-codec==1.1.7
Pillow==3.4.2
pinax-blog==5.0.2
pinax-cli==1.0.0
pinax-invitations==4.0.4
pinax-teams==0.12.1
pinax-theme-bootstrap==7.11.0
Pygments==2.1.3
python-openid==2.2.5
pytz==2016.7
requests==2.7.0
requests-oauthlib==0.7.0
simplejson==3.9.0
six==1.10.0
unicode-slugify==0.1.3
Unidecode==0.4.19
uritemplate==3.0.0
wheel==0.24.0
